requirements.txt
# Original library
jyotishganit>=0.1.2

# PyJHora library (alternative)
PyJHora>=4.5.0
pyswisseph>=2.10.0

# VedicAstro library (alternative)
VedicAstro>=0.2.0
# Note: Also install flatlib from sidereal branch:
# pip install git+https://github.com/diliprk/flatlib.git@sidereal

# Jyotishyamitra library (alternative)
Jyotishyamitra>=1.3.0

# Common dependencies
geopy>=2.4.0
pytz

# API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

